+++ app/renderer/package.json
@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@tanstack/react-virtual": "^3.0.4",
     "electron": "^28.1.4",
     "howler": "^2.2.4",
     "react": "^18.2.0",
