+++ app/renderer/src/pages/TrackList.tsx
@@ -1,9 +1,10 @@
-import React, { useContext, useEffect } from 'react';
+import React, { useContext, useEffect, useRef } from 'react';
 import { FaPlay } from 'react-icons/fa';
 import { FaPause } from 'react-icons/fa6';
 import { IoMdAdd } from 'react-icons/io';
 import { MdOutlineDownloadDone } from 'react-icons/md';
 import styled from 'styled-components';
+import { useVirtualizer } from '@tanstack/react-virtual';
 
 import { readDB, Track } from '../api';
 import { HorenContext } from '../App';
@@ -14,34 +15,45 @@ const TRACKLIST = styled.div`
   margin: 0;
   padding: 0;
   height: 100%;
-  overflow-y: auto;
-  // display: grid;
-  // padding-bottom: 88px;
-  // grid-template-columns: repeat(auto-fill, minmax(140px, 1fr));
-  table {
+  .track-list-container {
+    height: calc(100vh - 152px);
+    padding-right: 24px;
+  }
+  .list-header {
     color: #c6c6c6;
     width: 100%;
     border-collapse: collapse;
-  }
-  th {
     text-align: left;
-    font-size: 0.8rem;
-  }
-  tr {
-    &:nth-child(2n) {
-      background-color: #2c2c2c;
-    }
-    &:hover {
-      background-color: #3a3a3ab9;
+    display: flex;
+    div {
+      font-size: 0.8rem;
+      &:nth-child(1) {
+        margin: 0;
+      }
+      &:nth-child(2) {
+        flex-grow: 1;
+      }
+      &:nth-child(3) {
+        width: 100px;
+      }
+      &:nth-child(4) {
+        width: 100px;
+        margin: 0 16px;
+      }
     }
   }
 `;
 
-const PureItem = styled.tr`
+const PureItem = styled.div`
   color: #f1f1f1;
   margin: 4px 0;
+  padding-right: 8px;
   align-items: center;
   font-size: 0.9rem;
+  height: 100%;
+  &:nth-child(2n) {
+    background-color: #3a3a3a6d;
+  }
   &.playing {
     background-color: #10b45475 !important;
   }
@@ -51,12 +63,12 @@ const PureItem = styled.tr`
     margin-left: 4px;
   }
   .first {
-    width: 100%;
     display: flex;
     margin: 4px 0;
     align-items: center;
     justify-content: space-between;
     padding-right: 32px;
+    flex-grow: 1;
     .title {
       flex-grow: 1;
       min-width: 240px;
@@ -80,13 +92,26 @@ const PureItem = styled.tr`
   }
   .artist {
     font-weight: 300;
+    width: 120px;
+    text-align: left;
   }
   .album {
     font-weight: 300;
+    width: 160px;
+    padding: 0 8px;
+    text-align: left;
+  }
+  .duration {
+    text-align: center;
+    width: 40px;
   }
   .date {
-    min-width: 80px;
     text-align: center;
+    width: 60px;
+  }
+  .genre {
+    text-align: center;
+    width: 80px;
   }
 `;
 
@@ -99,6 +124,7 @@ export type TrackItemProps = {
   onPlayOrPause: (uid: string) => void;
   onAdd: (uid: string) => void;
   isAdd: (uid: string) => boolean;
+  style: React.CSSProperties;
 };
 
 export default function TrackList(props: PlayListPageProps) {
@@ -112,6 +138,14 @@ export default function TrackList(props: PlayListPageProps) {
     isPlaying,
     addToPlaylist,
   } = useContext(HorenContext);
+
+  const parentRef = useRef<HTMLTableElement | null>(null);
+
+  const rowVirtualizer = useVirtualizer({
+    count: trackList.length,
+    getScrollElement: () => parentRef.current,
+    estimateSize: () => 32,
+  });
 
   const handlePlayOrPause = (uid: string) => {
     playOrPause(uid);
@@ -130,35 +164,46 @@ export default function TrackList(props: PlayListPageProps) {
   return (
     <Page visible={visible}>
       <TRACKLIST
-        className="track-list perfect-scrollbar"
+        className="track-list"
         style={{ display: visible ? 'block' : 'none' }}
       >
-        <table>
-          <thead>
-            <tr>
-              <th style={{ textAlign: 'center' }}>序号</th>
-              <th>歌曲名</th>
-              <th>歌手</th>
-              <th>专辑</th>
-              <th>时长</th>
-              <th style={{ textAlign: 'center' }}>发行日期</th>
-              <th>流派</th>
-            </tr>
-          </thead>
-          <tbody>
-            {trackList?.map((track: Track, index: number) => (
-              <TrackPureItem
-                index={index + 1}
-                track={track}
-                onPlayOrPause={handlePlayOrPause}
-                onAdd={handleAdd}
-                key={track.src}
-                isAdd={isInPlaylist}
-                playing={isPlaying && current?.uid === track.uid}
-              />
-            ))}
-          </tbody>
-        </table>
+        <div
+          className="track-list-container perfect-scrollbar"
+          ref={parentRef}
+          style={{ overflow: 'auto', display: 'block' }}
+        >
+          <div
+            style={{
+              height: `${rowVirtualizer.getTotalSize()}px`,
+              width: '100%',
+              position: 'relative',
+            }}
+          >
+            {rowVirtualizer.getVirtualItems().map((virtualItem) => {
+              const track = trackList[virtualItem.index];
+              return (
+                <TrackPureItem
+                  key={track.src}
+                  style={{
+                    position: 'absolute',
+                    top: 0,
+                    left: 0,
+                    width: '100%',
+                    height: `${virtualItem.size}px`,
+                    transform: `translateY(${virtualItem.start}px)`,
+                    display: 'flex',
+                  }}
+                  index={virtualItem.index + 1}
+                  track={track}
+                  onPlayOrPause={handlePlayOrPause}
+                  onAdd={handleAdd}
+                  isAdd={isInPlaylist}
+                  playing={isPlaying && current?.uid === track.uid}
+                />
+              );
+            })}
+          </div>
+        </div>
       </TRACKLIST>
     </Page>
   );
@@ -171,6 +216,7 @@ function TrackPureItem({
   onPlayOrPause,
   onAdd,
   isAdd,
+  style,
 }: TrackItemProps) {
   const handlePlayOrPause = () => {
     if (onPlayOrPause && track) onPlayOrPause(track.uid);
@@ -181,54 +227,44 @@ function TrackPureItem({
   };
 
   return (
-    <PureItem key={track?.src} className={playing ? 'playing' : ''}>
-      <td>
-        <div className="index">
-          <span>{index}</span>
+    <PureItem
+      key={track?.src}
+      className={playing ? 'playing' : ''}
+      style={style}
+    >
+      <div className="index">
+        <span>{index}</span>
+      </div>
+      <div className="first">
+        <div className="title">{track?.title}</div>
+        <div className="operate">
+          {playing ? (
+            <div onClick={handlePlayOrPause} className="play">
+              <FaPause size={20} />
+            </div>
+          ) : (
+            <div onClick={handlePlayOrPause} className="play">
+              <FaPlay />
+            </div>
+          )}
+          {track && !isAdd(track.uid) ? (
+            <div onClick={handleAdd} className="add">
+              <IoMdAdd size={24} />
+            </div>
+          ) : (
+            <MdOutlineDownloadDone size={24} />
+          )}
         </div>
-      </td>
-      <td>
-        <div className="first">
-          <div className="title">{track?.title}</div>
-          <div className="operate">
-            {playing ? (
-              <div onClick={handlePlayOrPause} className="play">
-                <FaPause size={20} />
-              </div>
-            ) : (
-              <div onClick={handlePlayOrPause} className="play">
-                <FaPlay />
-              </div>
-            )}
-            {track && !isAdd(track.uid) ? (
-              <div onClick={handleAdd} className="add">
-                <IoMdAdd size={24} />
-              </div>
-            ) : (
-              <MdOutlineDownloadDone size={24} />
-            )}
-          </div>
-        </div>
-      </td>
-      <td>
-        <div className="artist">{track?.artist}</div>
-      </td>
-      <td>
-        <div className="album">{track?.album}</div>
-      </td>
-      <td>
-        <div className="duration">
-          {track?.duration !== undefined && normalizeDuration(track?.duration)}
-        </div>
-      </td>
-      <td>
-        <div className="date">
-          {track?.date !== 'undefined' ? track?.date?.slice(0, 4) : '未知'}
-        </div>
-      </td>
-      <td>
-        <div className="genre">{track?.genre}</div>
-      </td>
+      </div>
+      <div className="artist single-line">{track?.artist}</div>
+      <div className="album single-line">{track?.album}</div>
+      <div className="duration">
+        {track?.duration !== undefined && normalizeDuration(track?.duration)}
+      </div>
+      <div className="date">
+        {track?.date !== 'undefined' ? track?.date?.slice(0, 4) : '未知'}
+      </div>
+      <div className="genre">{track?.genre}</div>
     </PureItem>
   );
 }
